Moved timestamp generation to a utility function
