+++ frontend/src/components/Modals/styles.module.css
@@ -56,8 +56,6 @@
 }
 
 .askModalInput textarea {
-    resize: none;
-    font-family: Arial, Helvetica, sans-serif;
     height: 60px;
     border-color: #C71585;
     border-radius: 5px;
@@ -153,8 +151,6 @@
 }
 
 .offersModalInput textarea {
-    resize: none;
-    font-family: Arial, Helvetica, sans-serif;
     height: 60px;
     border-color: #3830a6;
     border-radius: 5px;
